more refactoring to get rid of constants
