+++ clients/android/app/src/main/java/app/lockbook/model/ListFilesViewModel.kt
@@ -6,6 +6,8 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.preference.PreferenceManager
 import androidx.work.WorkManager
+import app.lockbook.App.Companion.PERIODIC_SYNC_TAG
+import app.lockbook.App.Companion.UNEXPECTED_CLIENT_ERROR
 import app.lockbook.R
 import app.lockbook.ui.BreadCrumb
 import app.lockbook.ui.CreateFileInfo
@@ -14,9 +16,6 @@ import app.lockbook.ui.RenameFileInfo
 import app.lockbook.util.*
 import app.lockbook.util.FileMetadata
 import app.lockbook.util.FileType
-import app.lockbook.util.Messages.UNEXPECTED_CLIENT_ERROR
-import app.lockbook.util.RequestResultCodes.DRAWING_REQUEST_CODE
-import app.lockbook.util.RequestResultCodes.TEXT_EDITOR_REQUEST_CODE
 import app.lockbook.util.SharedPreferences.BACKGROUND_SYNC_ENABLED_KEY
 import app.lockbook.util.SharedPreferences.BACKGROUND_SYNC_PERIOD_KEY
 import app.lockbook.util.SharedPreferences.BIOMETRIC_OPTION_KEY
@@ -34,7 +33,6 @@ import app.lockbook.util.SharedPreferences.SORT_FILES_LAST_CHANGED
 import app.lockbook.util.SharedPreferences.SORT_FILES_TYPE
 import app.lockbook.util.SharedPreferences.SORT_FILES_Z_A
 import app.lockbook.util.SharedPreferences.SYNC_AUTOMATICALLY_KEY
-import app.lockbook.util.WorkManagerTags.PERIODIC_SYNC_TAG
 import com.beust.klaxon.Klaxon
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Ok
@@ -235,18 +233,10 @@ class ListFilesViewModel(path: String, application: Application) :
         }
     }
 
-    fun handleActivityResult(requestCode: Int) {
+    fun handleActivityResult() {
         uiScope.launch {
             withContext(Dispatchers.IO) {
-                when (requestCode) {
-                    TEXT_EDITOR_REQUEST_CODE, DRAWING_REQUEST_CODE -> {
-                        syncBasedOnPreferences()
-                    }
-                    else -> {
-                        Timber.e("Unable to recognize match requestCode: $requestCode.")
-                        _errorHasOccurred.postValue(UNEXPECTED_CLIENT_ERROR)
-                    }
-                }.exhaustive
+                syncBasedOnPreferences()
             }
         }
     }
