+++ clients/android/app/src/main/java/app/lockbook/ui/DrawingView.kt
@@ -62,6 +62,14 @@ class DrawingView(context: Context, attributeSet: AttributeSet?) :
         ERASER
     }
 
+    companion object {
+        const val CANVAS_WIDTH = 2125
+        const val CANVAS_HEIGHT = 2750
+
+        const val PRESSURE_SAMPLES_AVERAGED = 5
+        const val SPEN_ACTION_DOWN = 211
+    }
+
     private val scaleGestureDetector =
         ScaleGestureDetector(
             context,
